requirements.txt
bentoml
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
tqdm
dvc
fastapi[standard]
mlflow
torchview
ipykernel
pandas
numpy
seaborn
python-box
gdown
py-grpc-prometheus
# pip-chill
# wincertstore
# pytest
# tox
# flake8
# tox
# flake8
# mypy
# black
# e.